requirements.txt
Django==3.1
djangorestframework==3.12.4
django-environ==0.11.2
psycopg2-binary==2.9.10
pytest==7.1.3
pytest-django==4.9.0
pytest-cov==5.0.0
